fix de modulo soap mas vistas ok

+++ grpc-gateway/src/main/java/com/sistemas/distribuidos/grpc_gateway/dto/soap_externo/OngDTO.java
@@ -4,17 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import java.math.BigInteger;
-
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class OngDTO {
-    private BigInteger id;
+    private Long id;
     private String name;
     private String address;
     private String phone;
 }
-
